Removal of unused NPCs and addition of a custom doctor prefab
Removed "tailor", "botanist", "pharmacist" and "gunsmith" NPCPrefabs since they were never used anyways.

Added new NPC prefab "de_outpostdoctor" which is the same as the vanilla one, but instead of checking for a medical module and locking the NPC in place, it checks for a "doctor" spawnpoint tag to allow for custom placement.