Fixes to takeover events
Fixed missing SetDataAction which is required to reset meta data to acquire a new takeover mission.
Fixed coalition takeover missions trying to change ownership to separatists.
Changed order of event actions to allow the location ownership and metadata to change when mission succeeded, even if the event happens to get stuck on NPCChangeTeamAction or CombatAction, tho it was unlikely to ever happen, but better safe than sorry.