Fixed pirates attacking each other on pirate haven
Changed events to tag all hulls on the station beyond the security checkpoint with "piratehull" and to also tag all NPCs, including vanilla ones on the station with "piratenpc" so they all get targeted by NPCChangeTeamAction.

Added new event "piratecheckforweapon" which constantly checks every second for any actively held item with the tag "weapon" and if a piratenpc can actively see that item. If it succeeds, the entire outpost changes to a hostile team.