Biome gates open automatically if paid/high rep
People just can't comprehend manually flicking the lever.
The event will simply now set the lever's state to "true" when you got permission from the guy.
This still allows to break in and manually flick the lever.

Also removed dialog popup about the player having a weapon when a pirate on a pirate haven dies.