LocationTypes fixes and adjustments
Fixed "survey" not generating after initial map generation. It is now generated from "none" locations randomly, if there is no "mine" or "survey" within 2 jumps.

Adjusted "survey" to not change into a "mine" if there is a mine within 4 jumps.

Adjusted "survey" and "mine" to have a mincountperzone for the first biomes.

Reduced commonnessperzone for "none" locations.